+++ autopush/tests/test_integration.py
@@ -21,6 +21,7 @@ from twisted.internet.threads import deferToThread
 from twisted.web.client import Agent
 from twisted.test.proto_helpers import AccumulatingProtocol
 from autopush import __version__
+from base64 import urlsafe_b64encode
 
 log = logging.getLogger(__name__)
 here_dir = os.path.abspath(os.path.dirname(__file__))
@@ -161,7 +162,7 @@ keyid="http://example.org/bob/keys/123;salt="XZwpw6o37R-6qoZjw6KwAw"\
             method = "PUT"
 
         log.debug("%s body: %s", method, body)
-        http.request(method, url.path, body, headers)
+        http.request(method, url.path.encode("utf-8"), body, headers)
         resp = http.getresponse()
         http.close()
         log.debug("%s Response: %s", method, resp.read())
@@ -411,11 +412,65 @@ class TestSimple(IntegrationBase):
 
 class TestData(IntegrationBase):
     @inlineCallbacks
-    def test_data_delivery(self):
+    def test_simplepush_data_delivery(self):
         client = yield self.quick_register()
         result = yield client.send_notification(data="howdythere")
         ok_(result is not None)
         eq_(result["updates"][0]["data"], "howdythere")
+        yield self.shut_down(client)
+
+    @inlineCallbacks
+    def test_webpush_data_delivery_to_connected_client(self):
+        client = yield self.quick_register(use_webpush=True)
+        self.assertTrue(client.channels)
+        chan = client.channels.keys()[0]
+
+        # Invalid UTF-8 byte sequence.
+        data = b"\xc3\x28\xa0\xa1\xe2\x28\xa1"
+        result = yield client.send_notification(data=data)
+
+        ok_(result is not None)
+        eq_(result["messageType"], "notification")
+        eq_(result["channelID"], chan)
+        eq_(result["data"], "wyigoeIooQ==")
+
+        yield self.shut_down(client)
+
+    @inlineCallbacks
+    def test_webpush_data_delivery_to_disconnected_client(self):
+        tests = {
+            "d248d4e0-0ef4-41d9-8db5-2533ad8e4041": dict(
+                data=b"\xe2\x82\x28\xf0\x28\x8c\xbc", result="4oIo8CiMvA=="),
+
+            "df2363be-4d55-49c5-a1e3-aeae9450692e": dict(
+                data=b"\xf0\x90\x28\xbc\xf0\x28\x8c\x28",
+                result="8JAovPAojCg="),
+
+            "6c33e055-5762-47e5-b90c-90ad9bfe3f53": dict(
+                data=b"\xc3\x28\xa0\xa1\xe2\x28\xa1", result="wyigoeIooQ=="),
+        }
+
+        client = Client("ws://localhost:9010/", use_webpush=True)
+        yield client.connect()
+        yield client.hello()
+        for chan, test in tests.items():
+            yield client.register(chid=chan)
+
+        yield client.disconnect()
+        for chan, test in tests.items():
+            yield client.send_notification(channel=chan, data=test["data"])
+
+        yield client.connect()
+        yield client.hello()
+
+        for chan in tests:
+            result = yield client.get_notification()
+            ok_(result is not None)
+            chan = result["channelID"]
+            test = tests[chan]
+            eq_(result["data"], test["result"])
+            yield client.ack(chan, result["version"])
+
         yield self.shut_down(client)
 
     @inlineCallbacks
@@ -472,7 +527,7 @@ class TestWebPush(IntegrationBase):
         client = yield self.quick_register(use_webpush=True)
         result = yield client.send_notification(data=data)
         eq_(result["headers"]["encryption"], client._crypto_key)
-        eq_(result["data"], data)
+        eq_(result["data"], urlsafe_b64encode(data))
         eq_(result["messageType"], "notification")
         yield self.shut_down(client)
 
@@ -487,14 +542,14 @@ class TestWebPush(IntegrationBase):
         yield client.hello()
         result = yield client.get_notification()
         ok_(result != {})
-        eq_(result["data"], data)
+        eq_(result["data"], urlsafe_b64encode(data))
 
         yield client.disconnect()
         yield client.connect()
         yield client.hello()
         result = yield client.get_notification()
         ok_(result != {})
-        eq_(result["data"], data)
+        eq_(result["data"], urlsafe_b64encode(data))
         yield self.shut_down(client)
 
     @inlineCallbacks
@@ -510,20 +565,20 @@ class TestWebPush(IntegrationBase):
         yield client.hello()
         result = yield client.get_notification()
         ok_(result != {})
-        ok_(result["data"] in [data, data2])
+        ok_(result["data"] in map(urlsafe_b64encode, [data, data2]))
         result = yield client.get_notification()
         ok_(result != {})
-        ok_(result["data"] in [data, data2])
+        ok_(result["data"] in map(urlsafe_b64encode, [data, data2]))
 
         yield client.disconnect()
         yield client.connect()
         yield client.hello()
         result = yield client.get_notification()
         ok_(result != {})
-        ok_(result["data"] in [data, data2])
+        ok_(result["data"] in map(urlsafe_b64encode, [data, data2]))
         result = yield client.get_notification()
         ok_(result != {})
-        ok_(result["data"] in [data, data2])
+        ok_(result["data"] in map(urlsafe_b64encode, [data, data2]))
         yield self.shut_down(client)
 
     @inlineCallbacks
@@ -539,10 +594,10 @@ class TestWebPush(IntegrationBase):
         yield client.hello()
         result = yield client.get_notification()
         ok_(result != {})
-        ok_(result["data"] in [data, data2])
+        ok_(result["data"] in map(urlsafe_b64encode, [data, data2]))
         result = yield client.get_notification()
         ok_(result != {})
-        ok_(result["data"] in [data, data2])
+        ok_(result["data"] in map(urlsafe_b64encode, [data, data2]))
         yield client.ack(result["channelID"], result["version"])
 
         yield client.disconnect()
@@ -550,7 +605,7 @@ class TestWebPush(IntegrationBase):
         yield client.hello()
         result = yield client.get_notification()
         ok_(result != {})
-        ok_(result["data"] in [data, data2])
+        ok_(result["data"] in map(urlsafe_b64encode, [data, data2]))
         ok_(result["messageType"], "notification")
         result = yield client.get_notification()
         eq_(result, None)
@@ -569,10 +624,10 @@ class TestWebPush(IntegrationBase):
         yield client.hello()
         result = yield client.get_notification()
         ok_(result != {})
-        ok_(result["data"] in [data, data2])
+        ok_(result["data"] in map(urlsafe_b64encode, [data, data2]))
         result2 = yield client.get_notification()
         ok_(result2 != {})
-        ok_(result2["data"] in [data, data2])
+        ok_(result2["data"] in map(urlsafe_b64encode, [data, data2]))
         yield client.ack(result2["channelID"], result2["version"])
         yield client.ack(result["channelID"], result["version"])
 
@@ -595,7 +650,7 @@ class TestWebPush(IntegrationBase):
         yield client.hello()
         result = yield client.get_notification()
         eq_(result["channelID"], chan)
-        eq_(result["data"], data)
+        eq_(result["data"], urlsafe_b64encode(data))
 
         yield client.unregister(chan)
         yield client.disconnect()
@@ -612,7 +667,7 @@ class TestWebPush(IntegrationBase):
         client = yield self.quick_register(use_webpush=True)
         result = yield client.send_notification(data=data, ttl=0)
         eq_(result["headers"]["encryption"], client._crypto_key)
-        eq_(result["data"], data)
+        eq_(result["data"], urlsafe_b64encode(data))
         eq_(result["messageType"], "notification")
         yield self.shut_down(client)